requirements.txt
asgiref==3.5.2
Brotli==1.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.1
gcloud==0.18.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
jws==0.1.3
Naked==0.1.31
oauth2client==4.1.3
protobuf==4.21.7
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-decouple==3.6
python-jwt==2.0.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
whitenoise==6.2.0
